Rename validator list
